Better organize class package structure
