+++ src/main/java/org/bj/examples/trivia/service/slack/DelayedSlackService.java
@@ -1,4 +1,4 @@
-package org.bj.examples.trivia.service;
+package org.bj.examples.trivia.service.slack;
 
 import org.bj.examples.trivia.dto.SlackResponseDoc;
